initial divide of bs dashboard components
