+++ bookhive/src/components/bookStore/dashboard/InventorySummary.jsx
@@ -1,0 +1,9 @@
+
+
+const InventorySummary = () => {
+    return (
+    <>
+
+    </>);
+};
+export default InventorySummary;
