+++ bookhive/src/pages/bookstore/Dashboard.jsx
@@ -4,6 +4,8 @@ import {
   BookOpen, ExternalLink, Package, Truck, CheckCircle, Clock, Plus, Settings
 } from 'lucide-react';
 
+import RecentOrders from '../../components/bookStore/dashboard/RecentOrders';
+
 // MetricCard component inline
 const MetricCard = ({
   title,
@@ -46,130 +48,130 @@ const MetricCard = ({
 };
 
 // RecentOrders component inline
-const orders = [
-  {
-    id: 'BH-3842',
-    customer: 'Emma Thompson',
-    books: '3 items',
-    total: '$78.50',
-    status: 'delivered',
-    date: 'Jun 22, 2025'
-  },
-  {
-    id: 'BH-3841',
-    customer: 'Michael Chen',
-    books: '1 item',
-    total: '$24.99',
-    status: 'shipped',
-    date: 'Jun 21, 2025'
-  },
-  {
-    id: 'BH-3840',
-    customer: 'Sarah Johnson',
-    books: '2 items',
-    total: '$42.75',
-    status: 'processing',
-    date: 'Jun 20, 2025'
-  }
-];
+// const orders = [
+//   {
+//     id: 'BH-3842',
+//     customer: 'Emma Thompson',
+//     books: '3 items',
+//     total: '$78.50',
+//     status: 'delivered',
+//     date: 'Jun 22, 2025'
+//   },
+//   {
+//     id: 'BH-3841',
+//     customer: 'Michael Chen',
+//     books: '1 item',
+//     total: '$24.99',
+//     status: 'shipped',
+//     date: 'Jun 21, 2025'
+//   },
+//   {
+//     id: 'BH-3840',
+//     customer: 'Sarah Johnson',
+//     books: '2 items',
+//     total: '$42.75',
+//     status: 'processing',
+//     date: 'Jun 20, 2025'
+//   }
+// ];
 
-const getStatusIcon = (status) => {
-  switch (status) {
-    case 'delivered':
-      return <CheckCircle className="w-4 h-4 text-green-500" />;
-    case 'shipped':
-      return <Truck className="w-4 h-4 text-blue-500" />;
-    case 'processing':
-      return <Clock className="w-4 h-4 text-yellow-500" />;
-    default:
-      return <Package className="w-4 h-4 text-gray-400" />;
-  }
-};
+// const getStatusIcon = (status) => {
+//   switch (status) {
+//     case 'delivered':
+//       return <CheckCircle className="w-4 h-4 text-green-500" />;
+//     case 'shipped':
+//       return <Truck className="w-4 h-4 text-blue-500" />;
+//     case 'processing':
+//       return <Clock className="w-4 h-4 text-yellow-500" />;
+//     default:
+//       return <Package className="w-4 h-4 text-gray-400" />;
+//   }
+// };
 
-const getStatusColor = (status) => {
-  switch (status) {
-    case 'delivered':
-      return 'bg-green-50 text-green-700 border-green-200';
-    case 'shipped':
-      return 'bg-blue-50 text-blue-700 border-blue-200';
-    case 'processing':
-      return 'bg-yellow-50 text-yellow-700 border-yellow-200';
-    default:
-      return 'bg-gray-50 text-gray-700 border-gray-200';
-  }
-};
+// const getStatusColor = (status) => {
+//   switch (status) {
+//     case 'delivered':
+//       return 'bg-green-50 text-green-700 border-green-200';
+//     case 'shipped':
+//       return 'bg-blue-50 text-blue-700 border-blue-200';
+//     case 'processing':
+//       return 'bg-yellow-50 text-yellow-700 border-yellow-200';
+//     default:
+//       return 'bg-gray-50 text-gray-700 border-gray-200';
+//   }
+// };
 
-const RecentOrders = () => {
-  return (
-    <div className="bg-white rounded-xl shadow-sm border border-gray-100">
-      <div className="p-6 border-b border-gray-100">
-        <div className="flex items-center justify-between">
-          <h2 className="text-lg font-semibold text-gray-900">Recent Orders</h2>
-          <button className="flex items-center space-x-1 text-blue-600 hover:text-blue-700 text-sm font-medium">
-            <span>View all orders</span>
-            <ExternalLink className="w-4 h-4" />
-          </button>
-        </div>
-      </div>
+// const RecentOrders = () => {
+//   return (
+//     <div className="bg-white rounded-xl shadow-sm border border-gray-100">
+//       <div className="p-6 border-b border-gray-100">
+//         <div className="flex items-center justify-between">
+//           <h2 className="text-lg font-semibold text-gray-900">Recent Orders</h2>
+//           <button className="flex items-center space-x-1 text-blue-600 hover:text-blue-700 text-sm font-medium">
+//             <span>View all orders</span>
+//             <ExternalLink className="w-4 h-4" />
+//           </button>
+//         </div>
+//       </div>
       
-      <div className="overflow-x-auto">
-        <table className="w-full">
-          <thead className="bg-gray-50">
-            <tr>
-              <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
-                Order ID
-              </th>
-              <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
-                Customer
-              </th>
-              <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
-                Books
-              </th>
-              <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
-                Total
-              </th>
-              <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
-                Status
-              </th>
-              <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
-                Date
-              </th>
-            </tr>
-          </thead>
-          <tbody className="divide-y divide-gray-100">
-            {orders.map((order) => (
-              <tr key={order.id} className="hover:bg-gray-50 transition-colors duration-150">
-                <td className="py-4 px-6">
-                  <span className="text-sm font-medium text-blue-600">#{order.id}</span>
-                </td>
-                <td className="py-4 px-6">
-                  <span className="text-sm text-gray-900">{order.customer}</span>
-                </td>
-                <td className="py-4 px-6">
-                  <span className="text-sm text-gray-600">{order.books}</span>
-                </td>
-                <td className="py-4 px-6">
-                  <span className="text-sm font-medium text-gray-900">{order.total}</span>
-                </td>
-                <td className="py-4 px-6">
-                  <div className="flex items-center space-x-2">
-                    {getStatusIcon(order.status)}
-                    <span className={`inline-flex px-2 py-1 text-xs font-medium rounded-full border capitalize ${getStatusColor(order.status)}`}>
-                      {order.status}
-                    </span>
-                  </div>
-                </td>
-                <td className="py-4 px-6">
-                  <span className="text-sm text-gray-600">{order.date}</span>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  );
-};
+//       <div className="overflow-x-auto">
+//         <table className="w-full">
+//           <thead className="bg-gray-50">
+//             <tr>
+//               <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
+//                 Order ID
+//               </th>
+//               <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
+//                 Customer
+//               </th>
+//               <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
+//                 Books
+//               </th>
+//               <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
+//                 Total
+//               </th>
+//               <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
+//                 Status
+//               </th>
+//               <th className="text-left py-3 px-6 text-xs font-medium text-gray-500 uppercase tracking-wide">
+//                 Date
+//               </th>
+//             </tr>
+//           </thead>
+//           <tbody className="divide-y divide-gray-100">
+//             {orders.map((order) => (
+//               <tr key={order.id} className="hover:bg-gray-50 transition-colors duration-150">
+//                 <td className="py-4 px-6">
+//                   <span className="text-sm font-medium text-blue-600">#{order.id}</span>
+//                 </td>
+//                 <td className="py-4 px-6">
+//                   <span className="text-sm text-gray-900">{order.customer}</span>
+//                 </td>
+//                 <td className="py-4 px-6">
+//                   <span className="text-sm text-gray-600">{order.books}</span>
+//                 </td>
+//                 <td className="py-4 px-6">
+//                   <span className="text-sm font-medium text-gray-900">{order.total}</span>
+//                 </td>
+//                 <td className="py-4 px-6">
+//                   <div className="flex items-center space-x-2">
+//                     {getStatusIcon(order.status)}
+//                     <span className={`inline-flex px-2 py-1 text-xs font-medium rounded-full border capitalize ${getStatusColor(order.status)}`}>
+//                       {order.status}
+//                     </span>
+//                   </div>
+//                 </td>
+//                 <td className="py-4 px-6">
+//                   <span className="text-sm text-gray-600">{order.date}</span>
+//                 </td>
+//               </tr>
+//             ))}
+//           </tbody>
+//         </table>
+//       </div>
+//     </div>
+//   );
+// };
 
 // QuickActions component inline
 const ActionButton = ({ icon, title, description, onClick, color }) => (
@@ -278,7 +280,7 @@ const Dashboard = () => {
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
           <MetricCard
             title="Total Sales"
-            value="$4,278"
+            value="Yen4,278"
             change="+12% vs last month"
             isPositive={true}
             icon={DollarSign}
@@ -304,7 +306,7 @@ const Dashboard = () => {
             title="Avg. Order Value"
             value="$49.17"
             change="+1% vs last month"
-            isPositive={true}
+            isPositive={false}
             icon={TrendingUp}
             lastUpdated="Updated 2 hours ago"
           />
